requirements.txt
av==9.2.0
Pillow==9.2.0
python-tkdnd==0.2.1
python-vlc==3.0.16120
tk==0.1.0
tkvideoplayer==2.2
ttkwidgets==0.12.1
